add redirect using AMQP

+++ server/src/MessageStorage/TabTabGo.WebStream.MessageStorage/Services/StorageConnectionManager.cs
@@ -1,5 +1,6 @@
 ﻿using TabTabGo.Core.Data;
 using TabTabGo.WebStream.MessageStorage.Entites;
+using TabTabGo.WebStream.Model;
 using TabTabGo.WebStream.Services.Contract;
 
 namespace TabTabGo.WebStream.MessageStorage.Services
@@ -15,48 +16,50 @@ namespace TabTabGo.WebStream.MessageStorage.Services
             _userConnectionRepository = userConnectionRepository;
         }
 
-        public void RegisterConnection(string connectionId, string userId, IDictionary<string, object>? parameters = null)
+        public void RegisterConnection(string connectionId, UserIdData userId, IDictionary<string, object>? parameters = null)
         {
             _userConnectionRepository.Insert(new UserConnection
             {
                 ConnectionId = connectionId,
-                UserId = userId,
+                UserId = userId.UserId,
                 ExtraProperties = parameters, 
-                UpdatedBy = userId,
-                CreatedBy= userId,
+                UpdatedBy = userId.UserId,
+                CreatedBy= userId.UserId,
+                TenantId = userId.TenantId,
                 UpdatedDate = DateTime.Now
             });
             _unitOfWork.SaveChanges();
         }
 
-        public async Task RegisterConnectionAsync(string connectionId, string userId, IDictionary<string, object>? parameters = null, CancellationToken cancellationToken = default)
+        public async Task RegisterConnectionAsync(string connectionId, UserIdData userId, IDictionary<string, object>? parameters = null, CancellationToken cancellationToken = default)
         {
             await _userConnectionRepository.InsertAsync(new UserConnection
             {
                 ConnectionId = connectionId,
-                UserId = userId,
+                UserId = userId.UserId,
+                TenantId= userId.TenantId,
                 ExtraProperties = parameters,
-                UpdatedBy= userId,
+                UpdatedBy= userId.UserId,
                 UpdatedDate= DateTime.Now,
-                CreatedBy= userId,
+                CreatedBy= userId.UserId    ,
             }, cancellationToken);
             await _unitOfWork.SaveChangesAsync();
 
         }
-        public void UnRegisterConnection(string connectionId, string userId, IDictionary<string, object>? parameters = null)
+        public void UnRegisterConnection(string connectionId, UserIdData userId, IDictionary<string, object>? parameters = null)
         {
-            _userConnectionRepository.Delete(filter: userConn => userConn.ConnectionId == connectionId && userConn.UserId == userId);
+            _userConnectionRepository.Delete(filter: userConn => userConn.ConnectionId == connectionId && userConn.UserId == userId.UserId && userConn.TenantId==userId.TenantId);
             _unitOfWork.SaveChanges();
         }
 
-        public async Task UnRegisterConnectionAsync(string connectionId, string userId, IDictionary<string, object>? parameters = null, CancellationToken cancellationToken = default)
+        public async Task UnRegisterConnectionAsync(string connectionId, UserIdData userId, IDictionary<string, object>? parameters = null, CancellationToken cancellationToken = default)
         {
-            _userConnectionRepository.Delete(filter: userConn => userConn.ConnectionId == connectionId && userConn.UserId == userId);
+            _userConnectionRepository.Delete(filter: userConn => userConn.ConnectionId == connectionId && userConn.UserId == userId.UserId && userConn.TenantId==userId.TenantId);
             await _unitOfWork.SaveChangesAsync();
         }
-        public void ReRegisterConnection(string connectionId, string userId, IDictionary<string, object>? parameters = null)
+        public void ReRegisterConnection(string connectionId, UserIdData userId, IDictionary<string, object>? parameters = null)
         {
-            var userConnection = _userConnectionRepository.FirstOrDefault(filter: userConn => userConn.ConnectionId == connectionId && userConn.UserId == userId);
+            var userConnection = _userConnectionRepository.FirstOrDefault(filter: userConn => userConn.ConnectionId == connectionId && userConn.UserId == userId.UserId && userConn.TenantId==userId.TenantId);
             if (userConnection != null)
             {
                 userConnection.ReConnectedDate = DateTime.Now;
@@ -65,9 +68,9 @@ namespace TabTabGo.WebStream.MessageStorage.Services
             }
         }
 
-        public async Task ReRegisterConnectionAsync(string connectionId, string userId, IDictionary<string, object>? parameters = null, CancellationToken cancellationToken = default)
+        public async Task ReRegisterConnectionAsync(string connectionId, UserIdData userId, IDictionary<string, object>? parameters = null, CancellationToken cancellationToken = default)
         {
-            var userConnection = await _userConnectionRepository.FirstOrDefaultAsync(filter: userConn => userConn.ConnectionId == connectionId && userConn.UserId == userId, cancellationToken: cancellationToken);
+            var userConnection = await _userConnectionRepository.FirstOrDefaultAsync(filter: userConn => userConn.ConnectionId == connectionId && userConn.UserId == userId.UserId && userConn.TenantId == userId.TenantId, cancellationToken: cancellationToken);
             if (userConnection != null)
             {
                 userConnection.ReConnectedDate = DateTime.Now;
